+++ extensions/obsidian-tasks/src/hooks/useTasks.ts
@@ -21,6 +21,22 @@ export const useTasks = (preferences: Preferences) => {
       setTopTask(highestPriorityTask);
       const tasks = await getAllUncompletedTasks();
 
+      if (preferences.getOnlyCurrent) {
+        const today = new Date();
+        const tomorrow = new Date(today);
+        tomorrow.setDate(today.getDate() + 1);
+        tomorrow.setHours(0, 0, 0, 0);
+
+        for (let i = tasks.length - 1; i >= 0; i--) {
+          const dueDate = tasks[i].dueDate;
+          const scheduledDate = tasks[i].scheduledDate;
+
+          if ((dueDate && dueDate >= tomorrow) || (scheduledDate && scheduledDate >= tomorrow)) {
+            tasks.splice(i, 1);
+          }
+        }
+      }
+
       if (preferences.sortByPriority) {
         tasks.sort((a, b) => {
           const priorityA = a.priority || Priority.LOWEST;