+++ extensions/obsidian-tasks/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Obsidian Tasks Changelog
 
+## [Include only current tasks] - {PR_MERGE_DATE}
+
+- Add preference to include only current tasks
+
 ## [Initial Version] - 2025-04-11
 
 - List tasks with sorting and filtering